Add MultiThreading to DataBase

+++ src/main/java/ru/mail/polis/dao/senyast/model/Generation.java
@@ -12,7 +12,7 @@ public final class Generation {
 
     private Generation(){}
 
-    public static long fromPath(final Path path){
+    public static int fromPath(final Path path){
         return fromFileName(path.getFileName().toString());
     }
 
@@ -22,13 +22,13 @@ public final class Generation {
      * @param fileName name of FileTable file
      * @return generation number
      */
-    private static long fromFileName(final String fileName){
+    private static int fromFileName(final String fileName){
         final String pattern = LSMDao.PREFIX_FILE + "(\\d+)" + LSMDao.SUFFIX_DAT;
         final Pattern regex = Pattern.compile(pattern);
         final Matcher matcher = regex.matcher(fileName);
         if (matcher.find()){
-            return Long.parseLong(matcher.group(1));
+            return Integer.parseInt(matcher.group(1));
         }
-        return -1L;
+        return -1;
     }
 }
